requirements.txt
astroid==2.2.5
Click==7.0
Events==0.3
Flask==1.0.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
paho-mqtt==1.4.0
prometheus-client==0.7.1
protobuf==3.8.0
pylint==2.3.1
six==1.12.0
ttn==2.1.5
typed-ast==1.4.0
uWSGI==2.0.18
Werkzeug==0.15.4
wrapt==1.11.2
